Digital twinning
The purpose of this python program is to build a physical surrogate model using physical data to replicate a digital twin of a blended aircraft concept model. It then monitors, controls, and updates the model in order to optimize it for carbon neutrality, efficiency, and sustainability throughout the product lifecycle, and also creates reports and visualizations to track progress. The necessary libraries to execute this algorithm are numpy, tensorflow, and the tensorflow.keras layers library.